chore: Move to the utils package and remove lib.
