+++ RouteAnalyzer/src/test/kotlin/it/polito/group9/utils/UtilsTest.kt
@@ -1,4 +1,4 @@
-package it.polito.group9.lib
+package it.polito.group9.utils
 
 import it.polito.group9.model.WayPoint
 import org.junit.jupiter.api.Assertions
@@ -8,7 +8,7 @@ import kotlin.math.round
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
-class WaypointTest {
+class UtilsTest {
     @Tag("normalizeTimestamp")
     @Test
     fun normalizeTimestampShouldReturnTimestampWhenHasDecimal() {
@@ -67,6 +67,7 @@ class WaypointTest {
         assertEquals(expectedResult, result, "Result does not match expected result")
     }
 
+    @Tag("calculateEarthDistance")
     @Test
     fun `test calculateEarthDistance between known points`() {
         val p1 = WayPoint(0, 41.9028, 12.4964) // Roma
@@ -78,6 +79,7 @@ class WaypointTest {
         Assertions.assertEquals(1100.0, round(distance), 5.0)
     }
 
+    @Tag("calculateEarthDistance")
     @Test
     fun `test calculateEarthDistance with same point`() {
         val p = WayPoint(0, 45.0, 9.0)
@@ -88,6 +90,7 @@ class WaypointTest {
         Assertions.assertEquals(0.0, distance, 0.0001)
     }
 
+    @Tag("calculateEarthDistance")
     @Test
     fun `test calculateEarthDistance along meridian`() {
         val p1 = WayPoint(0, 0.0, 0.0)    // Equatore, longitudine 0
@@ -98,29 +101,4 @@ class WaypointTest {
         // 10 gradi di latitudine ~ 1110 km
         Assertions.assertEquals(1110.0, round(distance), 5.0)
     }
-
-    @Test
-    fun `test maxDistanceFromStart with multiple points`() {
-        val wayPoints = listOf(
-            WayPoint(0, 41.9028, 12.4964), // Roma (punto di partenza)
-            WayPoint(0, 48.8566, 2.3522),  // Parigi
-            WayPoint(0, 40.7128, -74.0060) // New York
-        )
-
-        val maxDistance = maxDistanceFromStart(wayPoints)
-
-        // New York è il punto più distante da Roma (~6889 km)
-        Assertions.assertEquals(6889.0, round(maxDistance), 5.0)
-    }
-
-    @Test
-    fun `test maxDistanceFromStart with single point`() {
-        val wayPoints = listOf(WayPoint(0, 41.9028, 12.4964)) // Solo Roma
-
-        val maxDistance = maxDistanceFromStart(wayPoints)
-
-        // Se c'è un solo WayPoint, la distanza massima è 0.0
-        Assertions.assertEquals(0.0, maxDistance, 0.0001)
-    }
-
 }
